requirement.txt
absl-py==0.11.0
astunparse==1.6.3
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
efficientnet-pytorch==0.7.0
fastapi==0.62.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h11==0.11.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
pkg-resources==0.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.7.3
python-dotenv==0.15.0
python-math==0.0.1
python-multipart==0.0.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
starlette==0.13.6
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
